video-to-text/requirements.txt
torch
torchvision
pillow
numpy
opencv-python
paho-mqtt
